+++ src/Map.hpp
@@ -8,12 +8,14 @@
 #ifndef MAP_HPP
 #define MAP_HPP
 
+#include <map>
 #include <memory>
 #include <set>
 #include <vector>
 
 namespace SLAM_demo {
 
+class Frame;
 class MapPoint;
 
 /**
@@ -28,10 +30,25 @@ public: // public members
     void addMPt(const std::shared_ptr<MapPoint>& pMPt);
     /// Get all map points from the map.
     std::vector<std::shared_ptr<MapPoint>> getAllMPts() const;
+    /// Get all related frames from the map.
+    std::vector<std::shared_ptr<Frame>> getAllFrames() const;
+    /** 
+     * @brief Update map point count data of a related frame, and remove 
+     *        redundant frame with no observed map point.
+     * @param[in] pFrame Pointer to a frame.
+     * @param[in] cnt    The map point count to be added to the related frame.
+     * @note @p pFrame will be removed from the map if its count becomes 0.
+     */
+    void updateFrameData(const std::shared_ptr<Frame>& pFrame, int cnt);
     /// Clear the current map.
     void clear();
     /// Remove redundant map points from the map.
     void removeMPts();
+    /** 
+     * @brief Transfer frames whose poses are optimized and will not receive 
+     *        further optimization to the System.
+     */
+    std::set<std::shared_ptr<Frame>> transferFramesOpt();
 private: // private data
     /**
      * @brief Minimum match-to-observation ratio (ratio of the number of times 
@@ -42,6 +59,15 @@ private: // private data
     static const unsigned TH_MAX_NUM_FRMS_LAST_SEEN;
     /// A set of map points.
     std::set<std::shared_ptr<MapPoint>> mspMPts;
+    /// A map of (related frame, number of observed map points) pairs.
+    std::map<std::shared_ptr<Frame>, int> mmRFrames;
+    /**
+     * @brief A set of frames whose poses are optimized.
+     * @note The frames are removed from the map, so no further optimization
+     *       is available. After the System fetches these frames, they will
+     *       be removed from the Map object.
+     */
+    std::set<std::shared_ptr<Frame>> mspFramesOpt;
 private: // private members
     /// Remove a point from the map.
     void removeMPt(std::shared_ptr<MapPoint>& pMPt);